refactor: moves client requests to another file
